API-WebService/requirements.txt
tensorflow==2.0.0b0
numpy==1.18.5
urllib3==1.25.9
Keras==2.3.1 
matplotlib==3.0.2
requests==2.23.0
gunicorn==20.0.4
Pillow==6.2.0
flask==1.1.2